Bam index construction - performance and other improvements.   Printed for code review

+++ src/java/net/sf/picard/util/BuildBamIndex.java
@@ -24,14 +24,14 @@
 
 package net.sf.picard.util;
 
-import net.sf.picard.cmdline.CommandLineProgram;
-import net.sf.picard.cmdline.Option;
-import net.sf.picard.cmdline.StandardOptionDefinitions;
-import net.sf.picard.cmdline.Usage;
+import net.sf.picard.cmdline.*;
 import net.sf.picard.io.IoUtil;
 import net.sf.samtools.*;
 
 import java.io.File;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.List;
 
 /**
  * Command line program to generate a BAM index (.bai) file from an existing BAM (.bam) file
@@ -42,18 +42,25 @@ public class BuildBamIndex extends CommandLineProgram {
     @Usage
     public String USAGE = getStandardUsagePreamble() + "Generates a BAM index (.bai) file. " +
             "Input BAM file must be sorted in coordinate order. " +
-            "Output file defaults to INPUT.bai (or x.bai if INPUT is x.bam)";
+            "Output file defaults to INPUT.bai (or x.bai if INPUT is x.bam)\n";
 
     @Option(shortName= StandardOptionDefinitions.INPUT_SHORT_NAME,
-            doc="A BAM file to process.")
+            doc="A BAM file to process.", optional=true, mutex="INPUT_URL")
     public File INPUT;
 
+    @Option(shortName= "URL",
+            doc="A BAM file to process.", optional=true, mutex="INPUT")
+    public String INPUT_URL;
+
     @Option(shortName=StandardOptionDefinitions.OUTPUT_SHORT_NAME,
-            doc="The BAM index file", optional=true)
+            doc="The BAM index file.", optional=true)
     public File OUTPUT;
 
-    @Option(doc = "Whether to sort the bins in bin number order")
+    @Option(doc = "Whether to sort the bins in bin number order.")
     public Boolean SORT = false;
+
+    @PositionalArguments(minElements = 0, maxElements = 2)
+    public List<File> IN_OUT;
 
     /** Stock main method for a command line program. */
     public static void main(final String[] argv) {
@@ -68,11 +75,24 @@ public class BuildBamIndex extends CommandLineProgram {
     protected int doWork() {
         final Log log = Log.getInstance(getClass());
 
-        // Some quick parameter checking
-        IoUtil.assertFileIsReadable(INPUT);
         IoUtil.assertFileIsWritable(OUTPUT);
+        final SAMFileReader bam;
 
-        final SAMFileReader bam = new SAMFileReader(INPUT);
+        if (INPUT_URL != null) {
+            // remote input
+            final URL bamURL;
+            try {
+                bamURL = new URL(INPUT_URL);
+            } catch (MalformedURLException e) {
+                throw new SAMException(e);
+            }
+            bam = new SAMFileReader(bamURL, null, false);
+
+        } else {
+            // input from a normal file
+            IoUtil.assertFileIsReadable(INPUT);
+            bam = new SAMFileReader(INPUT);
+        }
 
         if (!bam.isBinary()){
             throw new SAMException ("Input file must be bam file, not sam file.");
@@ -84,9 +104,9 @@ public class BuildBamIndex extends CommandLineProgram {
 
         bam.enableFileSource(true);
 
-        BAMIndexer instance = new BAMIndexer(INPUT, OUTPUT,
+        BAMIndexer instance = new BAMIndexer(OUTPUT,
                     bam.getFileHeader().getSequenceDictionary().size(), SORT);
-        instance.createIndex();
+        instance.createIndex(bam);
 
         log.info("Successfully wrote bam index file " + OUTPUT);
         return 0;
@@ -94,11 +114,28 @@ public class BuildBamIndex extends CommandLineProgram {
 
     @Override
     protected String[] customCommandLineValidation() {
-        if (OUTPUT == null){
+        // allow positional as alternative to i=input o=output
+        if (IN_OUT.size() > 0) {
+            if (INPUT != null) {
+                return new String[]{"Can't specify both named and positional INPUT parameter"};
+            }
+            INPUT = IN_OUT.get(0);
+            if (IN_OUT.size() > 1) {
+                if (OUTPUT != null) {
+                    return new String[]{"Can't specify both named and positional OUTPUT parameter"};
+                }
+                OUTPUT = IN_OUT.get(1);
+            }
+        }
+        if (INPUT == null && INPUT_URL == null)
+            return new String[]{"INPUT BAM file unspecified"};
+
+        // Todo allow output to a stream, e.g. BuildBamIndex > BaiToText
+        if (OUTPUT == null) {
             OUTPUT = new File(IoUtil.basename(INPUT) + BAMIndex.BAMIndexSuffix);
         }
-        if (OUTPUT.exists()){
-           OUTPUT.delete();
+        if (OUTPUT.exists()) {
+            OUTPUT.delete();
         }
         return null;
     }
